Add create authorised system service
Adding the service ready for use in  the controller. It has highlighted that 2 behaviours we want to prevent in the new version of the API are currently permitted

- do not allow duplicate client ID's
- do not allow authorised systems to have duplicate names

The first is covered in the old system because `client_id` is the ID in the table. But it does allow you to create multiple auth systems with the same name. As our 'users' are actually systems it's expected that we should not encounter any that have the same name. Plus there have been incidents of confusion caused by having 2 authorised system records both called 'wrls'. So in this new version we also want to make 'name' a unique field. We'll fix that in a subsequent change.