Use knex config to handle multiple environments
When we created this project as a replacement for the [charging-module-api](https://github.com/defra/charging-module-api) we followed some of its existing practises as we were just getting started. One of them was to hardcode which database to use into the some of the npm scripts.

Knex and specifically its config file `knexfile.js` has the ability to maintain a number of different configurations. Which is used is determined by what `NODE_ENV` is set to.

By switching to use the `knexfile.js` properly we are not only being more consistent with our knex usage, we are also enabling more flexibility by not having to hardcode database names into the project.