Drop dropdb commands
These were added in https://github.com/DEFRA/sroc-charging-module-api/pull/60 as part of the initial attempt to get this project working with Docker.

We can no longer get them to work. They appear to be postgres command line commands but without additional details we're not sure how to get them to work.