Update database config to use env vars only
With this change we can now drive our commands 100% through env vars.

We no longer need to hard code the db names. Instead we can define the db names as env vars. We can then use NODE_ENV to determine which db connection config to use when `knxefile.js` is required and used.

Should NODE_ENV not be set, we just default to `development` which will cover most use cases.