Fix createdb
Overlooked that createdb does not use `knexfile.js` as it has to be able to connect to the `postgres` db in order to drop our database. This meant it had to be tweaked to start using it to get connection information. We still don't pass it into knex when instantiating it though to allow us to connect to the 'postgres' db.