Implement bill run translator
Note. Currently we are validating the region directly in the bill run translator. It is something we are going to need to reuse when we come to add a transaction. But before we refactor we'd prefer t have implemented that other functionality so we have a clearer idea of how best to refactor the validator for reuse.