Rename bill run relationship
When we came to try and create a bill run we encountered an `Objection` error caused by the relationship name and the field used having the same name (once conversion to camelCase had been completed).

A simple fix is to just stick with the convention of naming the relationship after the thing it's relating too!