Remove BaseBillRunsController
This was here more as an example of the flexibility we have with our current model should we need to share functionality across controllers.

At this point though we only have support for a pre-sroc version of the bills runs controller so that is all we need to maintain at this point.