Pre-update the transactions controllers
https://trello.com/c/NFl1Q2PY

We are conducting a 'spike' on looking at how much creating and updating a customer summary at the time of adding a transaction effects performance. Our hope is the impact is negligible but doing so means we can drastically reduce the time it takes to generate a bill run summary.

We're doing this here under the `/v2` endpoints so that

- we don't effect any existing functionality
- we can be clear under which version of the endpoints the functionality has been implemented

This change is a first step in the spike, which is to tidy up our existing transaction controllers and the endpoints they support to match the features we need at this time.