Rename charge translator files with sroc prefix
The current functionality was written based on calculating an SROC charge for WRLS. The request coming in matches what was our current understanding of what would be needed to calculate an SROC WRLS charge.

With SROC shelved until 2022

- our understanding of what is needed for SROC is even more uncertain
- our priority is going to be replicating the current pre-sroc calculate charge endpoint
- we are not going to want to expose this endpoint for sometime

Also, to solve our problem with finding the right ruleset config we need to know whether we are dealing with SROC or not. So, if we rename the translator we can do 2 things

- make it clear this interpretation is based on an SROC charge
- we can set at the start of the process that this is an SROC request by setting a property on the translator

Hopefully, we can then keep the existing services and functionality generic ready for re-use when we come to implement a pre-sroc calculate charge endpoint.