Add a new ruleset field to the charge translator
The ruleset field is something we add within the translator to specify the ruleset that should be selected when querying the rules service.

By specifying it in the `_schema` we can not only default it. We can also allow it to be set in `data` at a later time should that become the preferred way.