Update ruleservice config to support multi rulesets
With this update to how the ruleservice config is held we can store multiple rulesets for each recognised regime. The key thing is whatever is using the config needs to know what ruleset to pull. Hence the changes to translator and presenter to bring `ruleset` as a property through to the `RulesService`.