Update calculate charge controller test
This was actually where we first started making changes. They started with first setting the tests up to actually work with the authorisation plugin. This meant ensuring we created a system user and recognised regime to go with it.

This then moved into making the outline of the test more consistent with others, for example, moving the database clean to the `beforeEach()` block rather than the `afterEach()` block.

But then we noticed some funnies. Though we were passing in a `fixture` value to `testFixtures()` we were never actually using it. When we spotted this we realised though all tests were named differently, because of the way they were written they were all running the eact same thing (`S126ProrataCreditFixtures`).

It's ideal if a fellow developer can look at a test in isolation and understand what is goin on. That's why we tend to make our tests even more explicit and verbose than the code.

So the final change was to move some of the test setup out of `testFixtures()` and back into the tests themselves even if it does lead to duplication.