Fix calculate charge dependence on unknown region (#77)
This was spotted when trying to resolve another issue. Currently, the authorisation plugin is expecting to pull `regimeId` from the path when a regime is set as a path param. This is then used to determine if the authenticating system has access to that regime.

The calculate charge endpoints specify `'/v2/{regime}/calculate-charge'` so the regime is not being picked up by the authentication plugin. We attempted to make the routes consistent and actually went for `regime` as the path param. This is because we are actually specifying the regime 'slug' (shortcode) rather than its ID so it makes better sense as a name.

So as part of this we were updating the `calculate_charge.controller.test`. We spotted within them we're making requests using a slug of `srocWrls`. Assuming this was a blip as there is no such regime currently seeded we changed this to `wrls`. At which point we broke everything 😳!

The reason is we're expecting the regime to match to ruleset config. So when we renamed the regime we broke this connection and the `RulesService` could no longer find the right config.

This change updates how the `RulesService` gets it's config so that

- it expects and can work with the currently known regimes
- it can handle moving between pre-sroc and sroc

We can then go back to making the regime path param consistent 😉!

* Additional Notes

1) We renamed the charge translator file with an `sroc` prefix. The current functionality was written based on calculating an SROC charge for WRLS. The request coming in matches our current understanding of what would be needed to calculate an SROC WRLS charge.

With SROC shelved until 2022

- our understanding of what is needed for SROC is even more uncertain
- our priority is going to be replicating the current pre-sroc calculate charge endpoint
- we are not going to want to expose this endpoint for sometime

Also, to solve our problem with finding the right ruleset config we need to know whether we are dealing with SROC or not. So, if we rename the translator we can do 2 things

- make it clear this interpretation is based on an SROC charge
- we can set at the start of the process that this is an SROC request by setting a property on the translator

Hopefully, we can then keep the existing services and functionality generic ready for re-use when we come to implement a pre-sroc calculate charge endpoint.

2) Add a new ruleset field to the charge translator

The ruleset field is something we add within the translator to specify the ruleset that should be selected when querying the rules service.

By specifying it in the `_schema` we can not only default it. We can also allow it to be set in `data` at a later time should that become the preferred way.

We also add ruleset as property to rules_service presenter so the `RulesService` can use it to grab the right config.

4) Update ruleservice config to support multi rulesets

With this update to how the ruleservice config is held we can now store multiple rulesets for each recognised regime. The key thing is whatever is using the config needs to know what ruleset to pull. Hence the changes to translator and presenter to bring `ruleset` as a property through to the `RulesService`.

5) Update calculate charge controller test

This was actually where we first started making changes. They started with first setting the tests to actually work with the authorisation plugin. This meant ensuring we created a system user and recognised regime to go with it.

This then moved into making the outline of the test more consistent with others, for example, moving the database clean to the `beforeEach()` block rather than the `afterEach()` block.

But then we noticed some funnies. Though we were passing in a `fixture` value to `testFixtures()` we were never actually using it. When we spotted this we realised although tests were named differently, because of the way they were written they were all running the exact same thing (`S126ProrataCreditFixtures`).

It's ideal if a fellow developer can look at a test in isolation and understand what is going on. That's why we tend to make our tests even more explicit and verbose than the code.

So the final change was to move some of the test set up out of `testFixtures()` and back into the tests themselves even if it does lead to duplication.

6) Update calculate charge controller to use the regime stored in `req.app`

The authorisation plugin will find and store an instance of `RegimeModel` in the `request.app` space. Again to be consistent, we grab our regime identifier from that now.