Apply feedback from review
Taking the feedback from the PR review we can make much better use of the spread operator to remove duplication in the config.

Also taken the point that by moving declaring the defaultConfig to the top of the new object it allows to override its properties should we choose to in the future.