Add create bill run service (#73)
This adds a new create bill run service ahead of adding a create bill run endpoint.

The new service will handle creating the bill run from the request and then return a result.

A key thing to highlight as we move forward with v2 of the API is we'll try to get out of the way of the db result and the API response. We have everything from the db, our result represents the full bill run record at that time, and when sent in the response it will be 100% valid JSON. There seems little point in then adding logic to meet the existing arbitrary pattern.