Use knex config to handle multiple environments (#70)
When we created this project as a replacement for the [charging-module-api](https://github.com/defra/charging-module-api) we followed some of its existing practises as we were just getting started. One of them was to hardcode which database to use into some of the npm scripts.

Knex and specifically its config file `knexfile.js` has the ability to maintain a number of different configurations. Which is used is determined by what `NODE_ENV` is set to.

By switching to use the `knexfile.js` properly we are not only being more consistent with our Knex usage, but we are also enabling more flexibility by not having to hardcode database names into the project.

Other changes

- Drop dropdb commands

These were added in https://github.com/DEFRA/sroc-charging-module-api/pull/60 as part of the initial attempt to get this project working with Docker.

We can no longer get them to work. They appear to be postgres command line commands but without additional details we're not sure how to get them to work.

- Fix createdb

Overlooked that createdb does not use `knexfile.js` as it has to be able to connect to the `postgres` db in order to drop our database. This meant it had to be tweaked to start using it to get connection information. We still don't pass it into Knex when instantiating it though to allow us to connect to the 'postgres' db.