Update bill runs ctrlr. to use new functionality (#74)
Now we have a proper bill run create service it's time to update our existing bill runs controllers to match how we expect them to operate.

At the moment we just have the create a bill run service. So this change updates the bill runs controllers to just support this functionality.