Add not-supported controller to project (#75)
This project is a replacement for the [charging-module-api](https://github.com/defra/chartging-module-api). It was originally intended as the means of implementing a new version of the charging rules and the requirements that go with it (SROC). But that has been postponed till 2022.

In the meantime, there are improvements to be made to the existing functionality, especially around the performance of generating the bill run. Plus we believe the design of this project is the only way we can support adding new regimes without duplicating all existing code. Duplication seems to be the only solution to adding regimes in the charging-module-api.

As these are large scale changes to the functionality we are doing them here so we at least have a 'working' version of the API for our first regime WRLS in the charging-module.

`/v2/` endpoints were going to represent SROC versions. Now we'll use them to represent updated versions of the existing endpoints. And as these are direct replacements there is no need to re-engineer the `/v1` endpoints as part of this project.

So to make it clear where we are not supporting a `/v1` version of an endpoint this change adds a new `NotSupportedController`.