Implement the translator
This covers all the validations found in the charging module API.

You will also see that there is duplication between it and the SROCChargeTranslator. This needs to be refactored out but as we are in the middle of a spike this is some 'tech debt' we're willing to stick on the credit card until the spike is complete. Then it will be task one to resolve.