Make docker dev environment more distinct (#82)
When working on the [charging-module-api Docker dev environment](https://github.com/DEFRA/charging-module-api/pull/175) we realised that the volume names used in the Docker compose were pretty generic. So resolved this in the charging-module by prefixing `cha_` to the names.

This change is just about solving the same issue in this project and ensuring we don't inadvertently overwrite or remove another service's volumes.