Migrate calculate charge from SROC to PRESROC
With the recent announcement that the proposed SROC charge scheme is postponsed until 2021/22 our previous work on implementing an SROC version of the calculate charge controller is not going to be used. It was also our 'best guess' at this time on what would be needed for an SROC calculate charge. With such a delay until the real thing, it is highly likely the implementation will need to change anyway.

But we do still need a PRESROC calculate charge and my migrating the functionality over it means it doesn't go to waste!

So these changes migrate the existing functionality for a calculate charge endpoint to work for PRESROC based requests instead 😀