Just return response body from ruleservice
Prior to this change we were always returning the full response we got from calling the rules service. We only actually cared about the body. To clean things up we now just return the body.

This also gives us a bit more flexibility in the service in how we call the rules service. For example, we are likely to want to put in a retry mechanism and inspect the messages it returns in future updates.