Reorganisation of the fixtures
Have rearranged to be by function first, then regime. The main reason for touching them was to update all the JSON fixtures to work for a presroc calculate charge endpoint. Moving and re-organising them in this way also helped during the process of ensuring we were using the right fixture.