Fix not handling unknown client id (#90)
https://trello.com/c/DWZta7a6

In preparation for a thorough test of our authentication and authorisation functionality, we did some exploratory testing ourselves. We found an issue where the service errors and returns a 500 when it should be returning a `401 - Unauthorized`.

If the token is valid (it is properly signed and not expired) but contains an unrecognised `client_id` we currently error.

This change fixes the issue. In this scenario, we now will return a `401 - Unauthorized`.