Migrate calculate charge from SROC to PRESROC (#87)
With the recent announcement that the proposed SROC charge scheme is postponed until 2021/22 our previous work on implementing an SROC version of the calculate charge controller is not going to be used. It was also our 'best guess' at this time on what would be needed for an SROC calculate charge. With such a delay until the real thing, it is highly likely the implementation will need to change anyway.

But we do still need a PRESROC calculate charge and by migrating the functionality over it means it doesn't go to waste!

So these changes migrate the existing functionality for a calculate charge endpoint to work for PRESROC based requests instead 😀

The work was pretty involved and covered some other key changes

* Implement a calculate charge service

We have a pattern of doing all 'work' in a service in order to allow us to unit test without having to make requests to a Hapi instance. So whilst refactoring the existing code to work with presroc we also moved it to a service and updated the controller to use it.

* Reorganisation of the fixtures

Rearranged to be by function first, then ruleset.

* Consistent naming

Translators, presenters etc have been renamed to `CalculateCharge` rather than just `Charge`

* Dropping ChargeModel

It was no longer needed after the changes made. We also are going to try to stick to a convention that classes in the `app/models` folder must have a representation on the database.

* Just return response body from ruleservice

Prior to this, we were always returning the full response we got from calling the rules service. We only actually care about the body. To clean things up we now just return the body.

This also gives us a bit more flexibility in the service class in how we call the rules service. For example, we are likely to want to put in a retry mechanism and inspect the messages it returns in future updates.

* Add a general test helper

We needed the ability to clone objects (fixtures) across some tests so this adds a `GeneralHelper` in testing.

* Fix the add bill run transaction ctrlr test

Not sure how this was passing before as it wasn't adding a user to the system!? Now it does and the test passes.