Only allow 'active' authorised systems
https://trello.com/c/mjGotliI

When we migrated the authorised system functionality over we added a status column. The intent is we can use this to control an authorised systems access without having to remove the record.

As yet we don't know what the different statuses will be. What we do know is that only authorised systems with a status of 'active' should be able to access the service.