Implement fix
Rather than use a pattern of extracting the status from the model and determining if its active or not, the calling code simply asks the model instance.

The reason we start the method name with a $ is to follow Objection's convention of prefixing all instance methods with $ to prevent them colliding with database column names. See https://vincit.github.io/objection.js/api/model/instance-methods.html