Only allow 'active' authorised systems (#92)
https://trello.com/c/mjGotliI

When we migrated the authorised system functionality over we added a status column. The intent is we can use this to control an authorised systems access without having to remove the record.

As yet we don't know what the different statuses will be. What we do know is that only authorised systems with a status of 'active' should be able to access the service.

* Note about the solution

Rather than use a pattern of extracting the status from the model and determining if it's active or not, the calling code simply asks the model instance.

The reason we start the method name with a $ is to follow Objection's convention of prefixing all instance methods with $ to prevent them from colliding with database column names. See https://vincit.github.io/objection.js/api/model/instance-methods.html