Update PostgreSQl dependence to v12
Currently the [charging-module-api](http://github.com/defra/charging-module-api) depends on version 10.6 of PostgreSQL. Our AWS environments are currently running that version and our original intent was to drop this version of the API in alongside until we are ready to switch.

Now we are working through so many changes as part of the performance rework with the intent of simply swapping out the old version we don't have that concern. So to take advantage of performance improvements in the latest version and to 'future-proof' the service for as long as possible we're going straight to PostgreSQL v12 when we deploy.

We won't have to change our postgres related packages are they are the latest already. But this change does update our development docker environment and CI to confirm we can work with v12.