Implement fix
The solution is based on trying to avoid directly referencing `NODE_ENV` in `server.js`. We have a `config/server.config.js` and that was felt to be the best fit for pulling in (and defaulting if not set) this env var.

As we previously just passed the whole object when starting Hapi this initially caused an error. Hence the tweak to the intent of `server.config.js`; it now represents the general config for our 'server' and includes the config needed for 'hapi'.

This isn't going to be the easiet of things to unit test so some local manual testing has confirmed the app is working as expected and we're happy that is sufficient.