Fix dev dependencies in production - again!
In [Fix reliance on dev-only dependencies in prod.](https://github.com/DEFRA/sroc-charging-module-api/pull/95) we thought we had solved the problem of not calling out to dev-only dependencies by not registering dev-only plugins.

The problem is the index.js file in `app/plugins` is still requiring them, which means they in turn are trying to `require()` missing packages in the production Docker image!

So this change includes a fix to prevent this from happening.