Override default hapi router options
This fixes an issue found after deploying the app to our AWS environments. Any requests to the service were returning a 404. When we checked the logs we found all the requests were being received with a trailing slash. For example, `GET https://myenv.cloud/status` was being recorded in the logs as `https://myenv.cloud/status/`.

This won't match to any of our routes hence the app returns a 404.

Some digging in the existing [charging-module-api](https://github.com/DEFRA/charging-module-api/blob/main/config/config.js#L27) confirmed it had made config changes to Hapi to deal with this situation. Our suspicion is that in the AWS API Gateway config for the environment something has inadvertently been set to add a trailing slash to the request before forwarding it to the app.

So the first change is to tell hapi to automatically strip any trailing slashes on incoming paths.

After discovering this router option we also spotted that Hapi is case sensitive by default. We want our app to respond to `/StAtuS` exactly the same way it responds to `/status`; a user shouldn't have to worry about case when making a request.

So the second change is we tell hapi to not be case senstive with incoming paths.