Add alter transaction table migration
When we create a transaction proper there is a host more detail we are expected to retain about it, which comes from the original request and the result of the rules service. There also looks to be a host of duplication and redundant fields!

But for now, we need to replicate the existing charging module and get things working. So part of that is updating our `transactions` table to match the existing [charging-module-api](https://github.com/defra/charging-module-api) one.

This change covers an alter table migration to do that.