Add transactions alter script
This adds _most_ of the remaining columns for the transaction table using the existing charging-module-api version as a basline. Some notable differences are

* Bill run centric

The current solution is a mish-mash of bill run and non-bill run based decisions. As all we know for sure now is WRLS is bill run centric there are some fields which would be better held at the bill run level. So these have been dropped (till we know more)

- transaction_file_id
- approved_for_billing
- approved_for_billing_at

* Never set

The following fields are either never set or changed from their defaults. This is based 264,293 transactions currently in our INTEGRATION environment, most of which are based on real data. So these have been dropped (till we know more)

- sequence_number
- related_reference
- currency_code
- header_narrative
- line_vat_code
- line_income_stream_code
- line_context_code
- line_quantity
- unit_of_measure

* Duplication

These fields are always set to the same value as `charge_value`. So these have been dropped (till we know more)

- currency_line_amount
- unit_of_measure_price

* Lack of flexibility

The current design has a boolean field called `pre_sroc`. The intent was to use it to decide if a transaction relates to the pre-sroc scheme. If false it would then be assumed it's for sroc. Our issue is this assumes there will only ever be 2 rulesets. With the uncertainty around SROC even this isn't a given. We don't know what the next ruleset will be, or if there will be more in the future.

So we have dropped this field in favour of capturing a `ruleset` name. This gives us much more flexibility whilst still being able to say a transaction is for 'presroc' or not.