Fix issue with authorised system migration scripts (#103)
We have found an issue with `db/migrations/20201127100927_alter_authorised_systems.js`. In it's drop it's trying to add the column again, not drop it. This change fixes the issue.

We also have realised we are not marking a boolean field as `notNullable()`. Ideally, all boolean fields should be defaulted and never contain a null so we never have to worry about whether a value is `true, false, or null`. So this also adds an alter script for the `authorised_systems` 'admin' field.