Add some tests but skip them
The main inspiration for these tests was taken from https://stackoverflow.com/a/40909092/6117745

However, the server module it is written against exposes its methods in such a way they are easy to stub/spy on. We've had no such luck with Hapi. So though we can't get our expectations to work we can see with well placed console logs that when these tests are run the behaviour we expect to see does happen.

We've left the tests here because

- it still serves as documentation of the behaviour we expect to happen with this feature
- we may solve the problem of stubbing and spying in the future at which point the tests can be fixed