Add stop (SIGTERM) plugin (#105)
Our deployment is based on Docker and [AWS ECS](https://aws.amazon.com/ecs/). When ECS needs to stop a container it will send a [SIGTERM](https://www.gnu.org/software/libc/manual/html_node/Termination-Signals.html) and allow it 30 seconds to stop.

> When UpdateService stops a task during a deployment, the equivalent of docker stop is issued to the containers running in the task. This results in a `SIGTERM` and a 30-second timeout, after which `SIGKILL` is sent and the containers are forcibly stopped. If the container handles the SIGTERM gracefully and exits within 30 seconds from receiving it, no `SIGKILL` is sent.
>
> https://awscli.amazonaws.com/v2/documentation/api/latest/reference/ecs/update-service.html

`SIGTERM` is a 'polite request' to stop. Typically it's caught by apps so they can do any clean up needed before shutting down. In the case of Hapi, it is often combined with [server.stop()](https://hapi.dev/api/?v=20.0.3#-await-serverstopoptions) to ensure no requests are dropped. When the signal is received an app will call `server.stop()` which tells Hapi to reject any new connections but allow any existing ones to complete. Without it, the app would be stopped immediately and any inflight requests would be dropped.

We want to avoid dropping requests if at all possible, so this change adds a new custom plugin to handle `SIGTERM` requests.

* Note. The solution is heavily based on [hapi-graceful-shutdown-plugin](https://github.com/visualjeff/hapi-graceful-shutdown-plugin). We've decided to 'do it ourselves' because

- the solution is relatively simple
- hapi-graceful-shutdown-plugin does not appear to be maintained. There are many outstanding Depdendabot requests
- we continue to try to strike a balance of only bringing in dependencies when we need them, whilst also not rewriting everything

* Note about skipped tests

The main inspiration for these tests was taken from https://stackoverflow.com/a/40909092/6117745

However, the server module it is written against exposes its methods in such a way they are easy to stub/spy on. We've had no such luck with Hapi. So though we can't get our expectations to work we can see with well placed console logs that when these tests are run the behaviour we expect to see does happen.

We've left the tests here because

- it still serves as documentation of the behaviour we expect to happen with this feature
- we may solve the problem of stubbing and spying in the future at which point the tests can be fixed