Allow access to validatedData in translators (#108)
https://trello.com/c/gTcQmR8M

When a `POST Transaction` request comes in we first pass it to the transaction translator. It has 2 jobs to do

- validate the transaction properties
- translate the properties to those that match our DB and model

Because of confusion on our part, we thought we could pass the translator directly to the `CalculateCharge` service, which would then treat the object as if it was the payload to the calculate charge endpoint. Not only would this keep things simple it would also mean we would not need to duplicate the validation done for the charge properties.

We overlooked that the transaction translator is converting something like `waterUndertaker` to `regimeValue14`. The first thing the calculate charge service does is pass in the payload to `CalculateChargeTranslator` which is expecting to see `waterUndertaker`. So everything blows up.

A simple solution to this which avoids the need for even more translators is to allow access to the validated data in the `TransactionTranslator`. If we can access that we can pass it as the payload to `CalculateChargeService` and then everything should work as expected.

This change stops hiding the data and exposes it using a new `validatedData` getter added to the `BaseTranslator` class.