Fix charge credit vs credit in transaction
When we were working on the spike in [PR #81](https://github.com/DEFRA/sroc-charging-module-api/pull/81) we had the intent of fixxing some of the transaction field names to make them simpler and remove unnecessary prefixes.

We've decided to hold off on those changes and focus on ensuring we have things behaving as they currently do first. Once of those changes that has managed to be implemented is referring to the `chargeCredit` field a `credit`.

To make things consistent with how the table has now been setup this small change rolls back the use of `credit` instead of `chargeCredit`.